Fix linting errors
Add the tests to the actual module
instead of it being in its own module
Add docstring
